python/django-postgres/pyproject.toml
[project]
name = "django-postgres-otel"
version = "1.0.0"
description = "Django REST API with OpenTelemetry instrumentation"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "base14", email = "hello@base14.io" }]
keywords = ["django", "opentelemetry", "observability", "rest-api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "django>=5.2,<5.3",
    "djangorestframework>=3.16,<4.0",
    "psycopg[binary]>=3.2",
    "dj-database-url>=2.3",
    "celery>=5.6,<6.0",
    "redis>=5.2,<6.0",
    "PyJWT>=2.10,<3.0",
    "python-dotenv>=1.0",
    "gunicorn>=23.0",
    "opentelemetry-sdk>=1.39",
    "opentelemetry-exporter-otlp>=1.39",
    "opentelemetry-instrumentation-django>=0.60b",
    "opentelemetry-instrumentation-psycopg>=0.59b",
    "opentelemetry-instrumentation-celery>=0.59b",
    "opentelemetry-instrumentation-redis>=0.59b",
    "opentelemetry-instrumentation-logging>=0.59b",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.14",
    "django-stubs>=5.1",
    "djangorestframework-stubs>=3.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "B904",   # raise without from in except
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = ["celery.*", "opentelemetry.*"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
addopts = "-v --tb=short"
